Move fragments tests
