+++ test/query_generator/fragments/fragment_glob_test.dart
@@ -1,7 +1,7 @@
 import 'package:artemis/generator/data.dart';
 import 'package:test/test.dart';
 
-import '../helpers.dart';
+import '../../helpers.dart';
 
 void main() {
   group('[Fragment generation]', () {
